requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-genai
googlesearch-python
ddgs
beautifulsoup4
curl-cffi
requests
